Extend fromInputDescriptor
fix testcases

Fix naming

Unify RequestOptions

Add test file

Enforce CLA

fix CHANGELOG.md

Release/4.1.1 (#105)

* 4.1.1: fix JSON serializer

* rename kmp-crypto references to signum

* easier migration fro renamed serializers
