+++ .github/workflows/cla.yml
@@ -1,0 +1,7 @@
+name: Verify CLA Acknowledgement
+on: [pull_request_target]
+
+jobs:
+  verify-cla:
+    uses: a-sit-plus/internal-workflows/.github/workflows/cla.yml@main
+
