+++ CONTRIBUTING.md
@@ -1,0 +1,70 @@
+<div align="center">
+<img src="https://github.com/user-attachments/assets/13750466-7d65-492f-9250-420da5467d2d">
+
+# Contributing to A-SIT Plus Open Source
+
+</div>
+
+We are happy to receive external contributions!
+"Just" opening issues for encountered problems is greatly appreciated too!
+
+## Reporting Issues
+
+We welcome constructive feedback and enquiries of any kind! Just be sure to check existing issues first to avoid duplicates.
+Please try to be as precise as possible and provide a reproducer (where applicable).
+
+## Contributing Changes
+
+We expect the contributor to hold all rights to the contributions they are about to commit.
+We particularly condemn copyright infringement and expect contributors to respect this position and strictly observe applicable law!
+
+If you plan on contributing changes to this repository's contents, please
+
+1. Fork it
+2. Create a branch with a descriptive name (e.g. `feature/timeTravel` or `fix/gravitationalConstant`)
+3. Commit your changes to your branch
+4. Open a pull request
+
+We will then review the changes, provide feedback.
+Once we agree that the PR is ready, we will approve it, and your PR will be merged.
+
+### Coding Conventions
+
+We try to follow the [official Kotlin coding conventions](https://kotlinlang.org/docs/coding-conventions.html) and expect the same from external contributors.
+We don't have any automated checks in place and hence don't strictly enforce convention rules by hard, but we will manually check for obvious violations.
+To put it plain and simple:
+
+> When in Rome, do as the Romans do!
+
+### Contributor Licence Agreement
+
+You must sign the [A-SIT Plus Contributor Licence Agreement](https://a-sit-plus.github.io/ASP_CLA.pdf) (CLA) before we can review your PR.
+This is a one-time requirement and applies to all A-SIT Plus open-source projects, which indicate to require external contributors to sign the CLA (such as this project).
+If you want to know more about what a CLA is and why some legal entities require contributors to sign it, check out the [Wikipedia page about CLAs](https://en.wikipedia.org/wiki/Contributor_License_Agreement).
+
+Please download and fill out the [A-SIT Plus CLA](https://a-sit-plus.github.io/ASP_CLA.pdf).
+Either print, sign, and scan it or (preferably) use an eIDAS-compliant qualified electronic signature.
+When done, send the signed CLA to opensource@a-sit.at.
+
+Information on the processing of your personal data can be found in the [A-SIT Plus Privacy Policy](https://a-sit-plus.github.io/ASP_Privacy.pdf).
+Be sure to carefully read it and only sign the CLA if you accept the terms of this privacy policy! (Don't worry it's a light read!)
+You will receive a confirmation message once we have received and processed your CLA.
+
+You are, of course, free to start your work prior to signing the CLA, but the review process can only start after we have received a signed copy of the CLA.
+We employ automated checks as part of our CI pipeline to verify external contributors have signed indeed the CLA.
+Accordingly, if you do open a PR prior to signing the CLA, the pipeline will fail and a publicly visible message will appear, tagging all contributors who still need to sign the CLA, including instructions on how to proceed.
+Once all contributors have signed the CLA, we will re-run the automated CLA checks.
+If all checks pass (i.e. all contributors have signed the CLA) we will start reviewing your contributions.
+
+If you have any questions on this matter don't hesitate to contact opensource@a-sit.at!
+
+<br>
+
+<hr>
+<div align="center">
+
+Copyright &copy; 2024 A-SIT Plus GmbH, Seidlgasse 22 / Top 9, 1030 Vienna, Austria
+
+<img src="https://a-sit-plus.github.io/logo_kurzertext.svg" width="163" height="64">
+
+</div>
