+++ gradle.properties
@@ -12,7 +12,7 @@ kotlin.mpp.enableCInteropCommonization=true
 kotlin.mpp.stability.nowarn=true
 kotlin.native.ignoreDisabledTargets=true
 
-artifactVersion = 4.2.0-SNAPSHOT
+artifactVersion = 4.1.1
 jdk.version=17
